Added -q --sequential option for running the program single-threaded.
This tool can be very memory intensive, enough so to cause the tool to fail to finish executing on less powerful hardware.  This modification allows the user to trade slower operation for less memory consumption, which is necessary for some uses.

In my case, this option takes about 3x longer to run, using about 10x less memory space.

Additionally, this is a suitable workaround for issue #7.